requirements.txt
pandas
numpy
fastapi[all]
scipy.sparse
implicit.als
pydantic